pyproject.toml
[tool.poetry]
name = "trading-bot"
version = "0.1.0"
description = "Simple trading strategy framework"
authors = ["Your Name <email@example.com>"]
packages = [{include = "trading_bot", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.0.0"
numpy = "^1.24.0"
ta = "^0.10.2"
binance-connector = "^3.7.0"
pymongo = "^4.5.0"
pydantic = "^2.0.0"
pyyaml = "^6.0"
tqdm = "^4.65.0"
deap = "^1.4.1"
scikit-learn = "^1.3.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
jupyter = "^1.0.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100